Codereview refactored spring api
