+++ Java/src/main/resources/application.properties
@@ -2,8 +2,8 @@ spring.datasource.url=jdbc:h2:mem:testdb
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=password
+spring.datasource.initialization-mode=always
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=create
-spring.datasource.initialization-mode=always
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
